# Conflicts:
#	app/src/main/java/com/apps/twelve/floor/salon/feature/catalog/adapters/GoodsListAdapter.java

+++ app/src/main/java/com/apps/twelve/floor/salon/feature/catalog/fragments/GoodsDetailsFragment.java
@@ -25,7 +25,6 @@ import com.arellomobile.mvp.presenter.InjectPresenter;
 import com.mzelzoghbi.zgallery.CustomViewPager;
 import com.mzelzoghbi.zgallery.adapters.HorizontalListAdapters;
 import java.util.ArrayList;
-import timber.log.Timber;
 
 /**
  * Created by Vrungel on 18.05.2017.
@@ -103,7 +102,6 @@ public class GoodsDetailsFragment extends BaseFragment implements IStaffDetailsF
       mViewPagerImages.setCurrentItem(currentPos);
 
       mCheckBoxFavoriteGoods.setChecked(mGoodsEntity.isFavorite());
-      Timber.e("" + mGoodsEntity.isFavorite());
 
       updateImageInfoAndButtons();
     }
@@ -147,10 +145,15 @@ public class GoodsDetailsFragment extends BaseFragment implements IStaffDetailsF
   }
 
   @OnClick(R.id.checkBoxFavoriteGoods) public void onCheckFavorite() {
-    if (mCheckBoxFavoriteGoods.isChecked()) {
-      mGoodsDetailsFragmentPresenter.addFavorite(mGoodsEntity.getId());
+    if (mAuthorizationManager.isAuthorized()) {
+      if (mCheckBoxFavoriteGoods.isChecked()) {
+        mGoodsDetailsFragmentPresenter.addFavorite(mGoodsEntity.getId());
+      } else {
+        mGoodsDetailsFragmentPresenter.deleteFavorite(mGoodsEntity.getId());
+      }
     } else {
-      mGoodsDetailsFragmentPresenter.deleteFavorite(mGoodsEntity.getId());
+      mCheckBoxFavoriteGoods.setChecked(false);
+      mGoodsDetailsFragmentPresenter.showAlertDialog();
     }
   }
